Move dereferencer to its own header
